refactor: add user rules file
